requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.7

# API Enhancement
django-filter==23.3
drf-spectacular==0.26.5

# Security
django-cors-headers==4.3.1

# Configuration
python-decouple==3.8

# Development
django-extensions>=3.2.3
django-debug-toolbar>=4.2.0

# Code Formatting & Linting
ruff>=0.1.0

